# contracts/openapi.yaml
...
openapi: 3.0.3
info:
  title: TaskFlow API
  version: 1.1.0
  description: A small multi-tenant task manager - login, logout, and CRUD over tasks.
servers:
  - url: http://127.0.0.1:8000
paths:
  /tenant:
    get:
      tags: [tenant]
      summary: Branding and enabled features for the customer owning this address
      description: >
        Which customer a request belongs to is decided from the Host header
        alone - never from a query parameter or a caller-supplied header,
        because the caller chooses those. An unrecognised host is a 404
        rather than a default, so a DNS mistake fails loudly instead of
        serving one customer's data under another's address.
      responses:
        "200":
          description: The current customer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantResponse" }
        "404":
          description: No customer is configured for this address
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /tasks/bulk-delete:
    post:
      tags: [tasks]
      summary: Delete all of the caller's finished tasks
      description: >
        An optional paid feature. Enforced on the server, not by hiding the
        button - a customer without the feature gets 403 even if they call
        the endpoint directly.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: How many tasks were removed
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted: { type: integer }
        "401":
          description: Not authenticated, or the token belongs to another customer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: This customer has not enabled bulk delete
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/login:
    post:
      tags: [auth]
      summary: Exchange credentials for a bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401":
          description: Credentials rejected. Identical for unknown user and wrong password.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke the presented token server-side
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Token revoked. Idempotent. }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /me:
    get:
      tags: [auth]
      summary: The authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /tasks:
    get:
      tags: [tasks]
      summary: List the caller's tasks, newest first
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status_filter
          in: query
          required: false
          schema: { $ref: "#/components/schemas/TaskStatus" }
        - name: q
          in: query
          required: false
          description: Case-insensitive substring match on the title
          schema: { type: string }
      responses:
        "200":
          description: Matching tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TaskResponse" }
    post:
      tags: [tasks]
      summary: Create a task. Always starts in `todo`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResponse" }
        "422":
          description: A domain rule rejected the input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /tasks/{task_id}:
    get:
      tags: [tasks]
      summary: Fetch one task
      security: [{ bearerAuth: [] }]
      parameters: [{ name: task_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: The task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResponse" }
        "404":
          description: Not found. Also returned for another user's task, deliberately.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      tags: [tasks]
      summary: Partially update a task
      security: [{ bearerAuth: [] }]
      parameters: [{ name: task_id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskUpdateRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Illegal status transition or invalid field
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      tags: [tasks]
      summary: Delete a task
      security: [{ bearerAuth: [] }]
      parameters: [{ name: task_id, in: path, required: true, schema: { type: string } }]
      responses:
        "204": { description: Deleted }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /test/users:
    post:
      tags: [test-support]
      summary: Test-data factory. 404 unless ENABLE_TEST_ENDPOINTS=true.
      parameters:
        - { name: username, in: query, required: false, schema: { type: string } }
        - { name: password, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Created user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "404":
          description: Test endpoints are disabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
    LoginRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 1, maxLength: 256 }
    TokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [bearer] }
    UserResponse:
      type: object
      required: [id, username, tenant_id]
      properties:
        id: { type: string }
        username: { type: string }
        tenant_id:
          type: string
          description: Which customer this user belongs to.
    TenantResponse:
      type: object
      required: [id, display_name, product_name, accent_colour, features]
      description: >
        A customer's public configuration. Mirrors tenants/<id>.yaml, and a
        certification test asserts the two agree - configuration that claims
        one thing while the code does another is the failure this guards.
      properties:
        id: { type: string }
        display_name: { type: string }
        product_name: { type: string }
        accent_colour: { type: string }
        features:
          type: array
          items: { type: string }
    TaskCreateRequest:
      type: object
      required: [title]
      additionalProperties: false
      properties:
        title: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 2000, nullable: true }
    TaskUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        title: { type: string, maxLength: 120, nullable: true }
        description: { type: string, maxLength: 2000, nullable: true }
        status:
          allOf: [{ $ref: "#/components/schemas/TaskStatus" }]
          nullable: true
    TaskResponse:
      type: object
      required: [id, owner_id, title, description, status, created_at, updated_at]
      properties:
        id: { type: string }
        owner_id: { type: string }
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ErrorResponse:
      type: object
      required: [code, detail]
      description: Every 4xx uses this shape. Branch on `code`, never on `detail`.
      properties:
        code: { type: string }
        detail: { type: string }